refactor: concern-split five near-cap AI lib modules (#4310 Slice B)
Extract build-planner input types, conquest-filter weight resolvers,
declare-war stalled/finalize bonuses, colonial-lite naval plan, and
recruitment candidate processing for durable headroom under the 400-line
source ceiling. No AI semantics change.

Refs #4310
